Slide and Carry
Added Slide effect to key before it is picked up. Added picking up of
the key. Key will follow the player around for now with no collider on
it. Next step is to make the player able to throw the key at which point
it should affected by all physics.